Broke apart model code into queries mutations and utilities.
