+++ meteor-app/imports/api/models/index.js
@@ -1,10 +1,24 @@
 // @flow
+export * from "./comment-mutations.js";
+export * from "./comment-queries.js";
 export * from "./comment.js";
-export * from "./misc.js";
+export * from "./company-mutations.js";
+export * from "./company-queries.js";
 export * from "./company.js";
+export * from "./job-ad-mutations.js";
+export * from "./job-ad-queries.js";
 export * from "./job-ad.js";
 export * from "./location.js";
+export * from "./misc.js";
+export * from "./review-mutations.js";
+export * from "./review-queries.js";
 export * from "./review.js";
+export * from "./salary-mutations.js";
+export * from "./salary-queries.js";
 export * from "./salary.js";
+export * from "./user-mutations.js";
+export * from "./user-queries.js";
 export * from "./user.js";
+export * from "./vote-mutations.js";
+export * from "./vote-queries.js";
 export * from "./vote.js";
